Move Color and Texture classes to separate files
