+++ renderer.cpp
@@ -26,61 +26,6 @@
 #include "gzstream.h"
 
 
-Color Color::WHITE(255, 255, 255);
-Color Color::YELLOW(250, 255, 0);
-Color Color::LIGHT_BROWN(210, 150, 0);
-Color Color::ORANGE_BROWN(250, 150, 0);
-Color Color::BROWN(240, 130, 0);
-Color Color::DARK_BROWN(100, 60, 0);
-Color Color::MAIN_MENU_ON(255, 255, 255, 190);
-Color Color::MAIN_MENU_OFF(255, 255, 255, 70);
-Color Color::LIGHT_GRAY(150, 150, 150);
-Color Color::GRAY(100, 100, 100);
-Color Color::ALMOST_GRAY(102, 102, 102);
-Color Color::DARK_GRAY(50, 50, 50);
-Color Color::ALMOST_BLACK(20, 20, 20);
-Color Color::ALMOST_DARK_GRAY(60, 60, 60);
-Color Color::BLACK(0, 0, 0);
-Color Color::ALMOST_WHITE(200, 200, 200);
-Color Color::GREEN(0, 255, 0);
-Color Color::LIGHT_GREEN(100, 255, 100);
-Color Color::DARK_GREEN(0, 150, 0);
-Color Color::RED(255, 0, 0);
-Color Color::LIGHT_RED(255, 100, 100);
-Color Color::PINK(255, 20, 147);
-Color Color::ORANGE(255, 165, 0);
-Color Color::BLUE(0, 0, 255);
-Color Color::NIGHT_BLUE(0, 0, 20);
-Color Color::DARK_BLUE(50, 50, 200);
-Color Color::LIGHT_BLUE(100, 100, 255);
-Color Color::SKY_BLUE(0, 191, 255);
-Color Color::PURPLE(160, 32, 240);
-Color Color::VIOLET(120, 0, 255);
-Color Color::TRANSLUCENT_BLACK(0, 0, 0);
-Color Color::TRANSPARENT(0, 0, 0, 0);
-
-Color::Color(Uint8 r, Uint8 g, Uint8 b, Uint8 a) : SDL_Color{r, g, b, a} {
-}
-
-Color Color::transparency(int trans) {
-  return Color(r, g, b, (Uint8)trans);
-}
-
-Color::Color() : Color(0, 0, 0) {
-}
-
-Color Color::f(double r, double g, double b, double a) {
-  return Color(r * 255, g * 255, b * 255, a * 255);
-}
-
-void Color::applyGl() const {
-  SDL::glColor4f((float)r / 255, (float)g / 255, (float)b / 255, (float)a / 255);
-}
-
-Color Color::operator* (Color c2) {
-  return Color(r * c2.r / 255, g * c2.g / 255, b * c2.b / 255, a * c2.a / 255);
-}
-
 Renderer::TileCoord::TileCoord(Vec2 p, int t) : pos(p), texNum(t) {
 }
 
@@ -90,155 +35,6 @@ Renderer::TileCoord::TileCoord() : TileCoord(Vec2(0, 0), -1) {
 static void checkOpenglError() {
   auto error = SDL::glGetError();
   CHECK(error == GL_NO_ERROR) << (int)error;
-}
-
-Texture::Texture(const FilePath& fileName) : path(fileName) {
-  SDL::SDL_Surface* image= SDL::IMG_Load(fileName.getPath());
-  CHECK(image) << SDL::IMG_GetError();
-  if (auto error = loadFromMaybe(image))
-    FATAL << "Couldn't load image: " << fileName << ". Error code " << toString(*error);
-  SDL::SDL_FreeSurface(image);
-}
-
-Texture::~Texture() {
-  if (texId)
-    SDL::glDeleteTextures(1, &(*texId));
-}
-
-static int totalTex = 0;
-
-// Some graphic cards need power of two sized textures, so we create a larger texture to contain the original one.
-SDL::SDL_Surface* Renderer::createPowerOfTwoSurface(SDL::SDL_Surface* image) {
-  int w = 1;
-  int h = 1;
-  while (w < image->w)
-    w *= 2;
-  while (h < image->h)
-    h *= 2;
-  if (w == image->w && h == image->h)
-    return image;
-  auto ret = createSurface(w, h);
-  SDL::SDL_Rect dst {0, 0, image->w, image->h };
-  SDL::SDL_SetSurfaceBlendMode(image, SDL::SDL_BLENDMODE_NONE);
-  SDL_BlitSurface(image, nullptr, ret, &dst);
-  // fill the rest of the texture as well, which 'kind-of' solves the problem with repeating textures.
-  dst = {image->w, 0, 0, 0};
-  SDL_BlitSurface(image, nullptr, ret, &dst);
-  dst = {image->w, image->h, 0, 0};
-  SDL_BlitSurface(image, nullptr, ret, &dst);
-  dst = {0, image->h, 0, 0};
-  SDL_BlitSurface(image, nullptr, ret, &dst);
-  return ret;
-}
-
-optional<SDL::GLenum> Texture::loadFromMaybe(SDL::SDL_Surface* imageOrig) {
-  if (!texId) {
-    texId = 0;
-    SDL::glGenTextures(1, &(*texId));
-  }
-  checkOpenglError();
-  SDL::glBindTexture(GL_TEXTURE_2D, (*texId));
-  checkOpenglError();
-  SDL::glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-  SDL::glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-  SDL::glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT);
-  SDL::glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
-  checkOpenglError();
-  int mode = GL_RGB;
-  auto image = Renderer::createPowerOfTwoSurface(imageOrig);
-  if (image->format->BytesPerPixel == 4) {
-    if (image->format->Rmask == 0x000000ff)
-      mode = GL_RGBA;
-    else
-      mode = GL_BGRA;
-  } else {
-    if (image->format->Rmask == 0x000000ff)
-      mode = GL_RGB;
-    else
-      mode = GL_BGR;
-  }
-  SDL::glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
-  SDL::glPixelStorei(GL_UNPACK_ROW_LENGTH, 0);
-  SDL::glPixelStorei(GL_UNPACK_SKIP_PIXELS, 0);
-  SDL::glPixelStorei(GL_UNPACK_SKIP_ROWS, 0);
-  checkOpenglError();
-  SDL::glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, image->w, image ->h, 0, mode, GL_UNSIGNED_BYTE, image->pixels);
-  size = Vec2(imageOrig->w, imageOrig->h);
-  realSize = Vec2(image->w, image->h);
-  if (image != imageOrig)
-    SDL::SDL_FreeSurface(image);
-  auto error = SDL::glGetError();
-  if (error != GL_NO_ERROR)
-    return error;
-  return none;
-}
-
-Texture::Texture(const FilePath& filename, int px, int py, int w, int h) : path(filename) {
-  SDL::SDL_Surface* image = SDL::IMG_Load(path->getPath());
-  CHECK(image) << SDL::IMG_GetError();
-  SDL::SDL_Rect offset;
-  offset.x = 0;
-  offset.y = 0;
-  SDL::SDL_Rect src { px, py, w, h };
-  SDL::SDL_Surface* sub = Renderer::createSurface(src.w, src.h);
-  CHECK(sub) << SDL::SDL_GetError();
-  CHECK(!SDL_BlitSurface(image, &src, sub, &offset)) << SDL::SDL_GetError();
-  SDL::SDL_FreeSurface(image);
-  if (auto error = loadFromMaybe(sub))
-    FATAL << "Couldn't load image: " << *path << ". Error code " << toString(*error);
-  SDL::SDL_FreeSurface(sub);
-}
-
-Texture::Texture(Color color, int width, int height) {
-  vector<Color> colors(width * height, color);
-  texId = 0;
-  SDL::glGenTextures(1, &(*texId));
-  checkOpenglError();
-  SDL::glBindTexture(GL_TEXTURE_2D, (*texId));
-  SDL::glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, colors.data());
-  realSize = size = Vec2(width, height);
-  SDL::glBindTexture(GL_TEXTURE_2D, 0);
-}
-
-Texture::Texture(SDL::SDL_Surface* surface) {
-  CHECK(!loadFromMaybe(surface));
-}
-
-Texture::Texture(Texture&& tex) {
-  *this = std::move(tex);
-}
-
-Texture& Texture::operator = (Texture&& tex) {
-  size = tex.size;
-  realSize = tex.realSize;
-  texId = tex.texId;
-  path = tex.path;
-  tex.texId = none;
-  return *this;
-}
-
-optional<Texture> Texture::loadMaybe(const FilePath& path) {
-  if (SDL::SDL_Surface* image = SDL::IMG_Load(path.getPath())) {
-    Texture ret;
-    bool ok = !ret.loadFromMaybe(image);
-    SDL::SDL_FreeSurface(image);
-    if (ok) {
-      ret.path = path;
-      return std::move(ret);
-    }
-  }
-  return none;
-}
-
-const Vec2& Texture::getSize() const {
-  return size;
-}
-
-void Texture::addTexCoord(int x, int y) const {
-  SDL::glTexCoord2f((float)x / realSize.x, (float)y / realSize.y);
-}
-
-Texture::Texture() {
 }
 
 void Renderer::renderDeferredSprites() {
@@ -258,7 +54,7 @@ void Renderer::renderDeferredSprites() {
     colors.push_back(((float) color.b) / 255);
     colors.push_back(((float) color.a) / 255);
   };
-  optional<SDL::GLuint> currentTex;
+  SDL::GLuint currentTex = 0;
   for (auto& elem : deferredSprites) {
     auto add = [&](Vec2 v, int texX, int texY, const DeferredSprite& draw) {
       addVertex(v, texX, texY, draw.realSize, draw.color.value_or(Color::WHITE));
@@ -271,7 +67,7 @@ void Renderer::renderDeferredSprites() {
     add(elem.d, elem.p.x, elem.k.y, elem);
   }
   if (!vertices.empty()) {
-    SDL::glBindTexture(GL_TEXTURE_2D, *currentTexture);
+    SDL::glBindTexture(GL_TEXTURE_2D, currentTexture);
     SDL::glEnable(GL_TEXTURE_2D);
     checkOpenglError();
     SDL::glEnableClientState(GL_VERTEX_ARRAY);
@@ -303,10 +99,10 @@ void Renderer::drawSprite(const Texture& t, Vec2 topLeft, Vec2 bottomRight, Vec2
 }
 
 void Renderer::drawSprite(const Texture& t, Vec2 a, Vec2 b, Vec2 c, Vec2 d, Vec2 p, Vec2 k, optional<Color> color) {
-  if (currentTexture && currentTexture != *t.texId)
+  if (currentTexture && currentTexture != t.getTexId())
     renderDeferredSprites();
-  currentTexture = *t.texId;
-  deferredSprites.push_back({a, b, c, d, p, k, t.realSize, color});
+  currentTexture = t.getTexId();
+  deferredSprites.push_back({a, b, c, d, p, k, t.getRealSize(), color});
 }
 
 static float sizeConv(int size) {
@@ -581,7 +377,7 @@ SDL::SDL_Surface* Renderer::loadScaledSurface(const FilePath& path, double scale
   if (auto surface = SDL::IMG_Load(path.getPath())) {
     if (scale == 1)
       return surface;
-    if (auto scaled = createSurface(surface->w * scale, surface->h * scale)) {
+    if (auto scaled = Texture::createSurface(surface->w * scale, surface->h * scale)) {
       SDL::SDL_SetSurfaceBlendMode(surface, SDL::SDL_BLENDMODE_NONE);
       CHECK(!SDL_BlitScaled(surface, nullptr, scaled, nullptr)) << SDL::IMG_GetError();
       SDL_FreeSurface(surface);
@@ -752,17 +548,11 @@ void Renderer::setAnimationsDirectory(const DirectoryPath& path) {
   animationDirectory = path;
 }
 
-SDL::SDL_Surface* Renderer::createSurface(int w, int h) {
-  SDL::SDL_Surface* ret = SDL::SDL_CreateRGBSurface(0, w, h, 32, 0x000000FF, 0x0000FF00, 0x00FF0000, 0xFF000000);
-  CHECK(ret) << "Failed to creature surface " << w << ":" << h << ": " << SDL::SDL_GetError();
-  return ret;
-}
-
 void Renderer::loadTilesFromDir(const DirectoryPath& path, Vec2 size, int setWidth) {
   const static string imageSuf = ".png";
   auto files = path.getFiles().filter([](const FilePath& f) { return f.hasSuffix(imageSuf);});
   int rowLength = setWidth / size.x;
-  SDL::SDL_Surface* image = createSurface(setWidth, setWidth);
+  SDL::SDL_Surface* image = Texture::createSurface(setWidth, setWidth);
   SDL::SDL_SetSurfaceBlendMode(image, SDL::SDL_BLENDMODE_NONE);
   CHECK(image) << SDL::SDL_GetError();
   int frameCount = 0;
