+++ renderer.h
@@ -20,52 +20,8 @@
 #include "util.h"
 #include "file_path.h"
 #include "animation_id.h"
-
-namespace fx {
-class FXRenderer;
-}
-
-struct Color : public SDL::SDL_Color {
-  Color(Uint8, Uint8, Uint8, Uint8 = 255);
-  Color transparency(int);
-  static Color f(double, double, double, double = 1.0);
-  Color operator* (Color);
-  Color();
-  void applyGl() const;
-
-  static Color WHITE;
-  static Color MAIN_MENU_ON;
-  static Color MAIN_MENU_OFF;
-  static Color YELLOW;
-  static Color LIGHT_BROWN;
-  static Color ORANGE_BROWN;
-  static Color BROWN;
-  static Color DARK_BROWN;
-  static Color LIGHT_GRAY;
-  static Color GRAY;
-  static Color ALMOST_GRAY;
-  static Color DARK_GRAY;
-  static Color ALMOST_BLACK;
-  static Color ALMOST_DARK_GRAY;
-  static Color BLACK;
-  static Color ALMOST_WHITE;
-  static Color GREEN;
-  static Color LIGHT_GREEN;
-  static Color DARK_GREEN;
-  static Color RED;
-  static Color LIGHT_RED;
-  static Color PINK;
-  static Color ORANGE;
-  static Color BLUE;
-  static Color DARK_BLUE;
-  static Color NIGHT_BLUE;
-  static Color LIGHT_BLUE;
-  static Color SKY_BLUE;
-  static Color PURPLE;
-  static Color VIOLET;
-  static Color TRANSLUCENT_BLACK;
-  static Color TRANSPARENT;
-};
+#include "color.h"
+#include "texture.h"
 
 enum class SpriteId {
   BUILDINGS,
@@ -80,35 +36,6 @@ class ViewObject;
 class Clock;
 
 struct sth_stash;
-
-class Texture {
-  public:
-  Texture();
-  Texture(const Texture&) = delete;
-  Texture(Texture&&);
-  Texture& operator = (Texture&&);
-  Texture(const FilePath& path);
-  Texture(const FilePath& path, int px, int py, int kx, int ky);
-  Texture(Color, int width, int height);
-
-  explicit Texture(SDL::SDL_Surface*);
-  static optional<Texture> loadMaybe(const FilePath&);
-
-  optional<SDL::GLenum> loadFromMaybe(SDL::SDL_Surface*);
-  const Vec2& getSize() const;
-
-  ~Texture();
-
-  private:
-  friend class Renderer;
-  friend class fx::FXRenderer;
-
-  void addTexCoord(int x, int y) const;
-  optional<SDL::GLuint> texId;
-  Vec2 size;
-  Vec2 realSize;
-  optional<FilePath> path;
-};
 
 class Renderer {
   public:
@@ -140,8 +67,6 @@ class Renderer {
   static vector<string> getFullscreenResolutions();
   const static int textSize = 19;
   const static int smallTextSize = 14;
-  static SDL::SDL_Surface* createSurface(int w, int h);
-  static SDL::SDL_Surface* createPowerOfTwoSurface(SDL::SDL_Surface*);
   enum FontId { TEXT_FONT, TILE_FONT, SYMBOL_FONT };
   int getTextLength(const string& s, int size = textSize, FontId = TEXT_FONT);
   Vec2 getTextSize(const string& s, int size = textSize, FontId = TEXT_FONT);
@@ -251,7 +176,7 @@ class Renderer {
   SDL::SDL_Cursor* cursor;
   SDL::SDL_Cursor* cursorClicked;
   SDL::SDL_Surface* loadScaledSurface(const FilePath& path, double scale);
-  optional<SDL::GLuint> currentTexture;
+  SDL::GLuint currentTexture = 0;
   void drawSprite(const Texture& t, Vec2 a, Vec2 b, Vec2 c, Vec2 d, Vec2 p, Vec2 k, optional<Color> color);
   void drawSprite(const Texture& t, Vec2 topLeft, Vec2 bottomRight, Vec2 p, Vec2 k, optional<Color> color);
   struct DeferredSprite {
